refactor : 파일 구조 변경 도메인 중심 패키징 형식으로 변경
