+++ src/main/java/com/hyun/oauthboard/domain/board/entity/Board.java
@@ -1,8 +1,6 @@
 package com.hyun.oauthboard.domain.board.entity;
 
 import com.hyun.oauthboard.domain.board.dto.BoardCreateRequestDto;
-import com.hyun.oauthboard.domain.board.dto.BoardResponse;
-import com.hyun.oauthboard.domain.board.entity.model.BoardEntityModel;
 import com.hyun.oauthboard.domain.member.entity.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -12,18 +10,21 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
 
 @Builder
 @Entity
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Board extends BoardEntityModel {
+public class Board {
 
     @Id
     @Column(name = "board_id")
@@ -40,19 +41,16 @@ public class Board extends BoardEntityModel {
     @Column(name = "board_content")
     private String boardContent;
 
+    @CreatedDate
+    @Column(updatable = false, name = "created_at")
+    private LocalDateTime createdAt;
+
+    @LastModifiedDate
+    @Column(name = "updated_at")
+    private LocalDateTime updatedAt;
+
     public void updateBoard(BoardCreateRequestDto boardCreateRequestDto) {
         this.boardTitle = boardCreateRequestDto.getBoardTitle();
         this.boardContent = boardCreateRequestDto.getBoardContent();
-    }
-
-    public BoardResponse toDto() {
-        return BoardResponse.builder()
-            .boardId(this.boardId)
-            .memberId(this.getMember().getMemberId())
-            .boardTitle(this.boardTitle)
-            .boardContent(this.boardContent)
-            .createdAt(this.getCreatedAt())
-            .updatedAt(this.getUpdatedAt())
-            .build();
     }
 }
