+++ src/main/java/com/hyun/oauthboard/domain/boardLike/controller/BoardLikeController.java
@@ -1,0 +1,10 @@
+package com.hyun.oauthboard.domain.boardLike.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/board/like")
+public class BoardLikeController {
+
+}
