+++ src/main/java/com/hyun/oauthboard/domain/boardLike/repository/BoardLikeRepository.java
@@ -1,0 +1,8 @@
+package com.hyun.oauthboard.domain.boardLike.repository;
+
+import com.hyun.oauthboard.domain.boardLike.entity.BoardLike;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BoardLikeRepository extends JpaRepository<BoardLike, Long> {
+
+}
